requirements.txt
aiogram
sqlalchemy[asyncio]
asyncpg
python-dotenv
alembic
pydantic
pydantic-settings
emoji
pydantic
redis